backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
python-multipart==0.0.6
google-generativeai==0.3.2
opencv-python==4.9.0.80
pytesseract==0.3.10
pillow==10.2.0
numpy==1.26.3
duckduckgo-search==4.2
wikipedia-api==0.6.0
torch==2.1.2
transformers==4.36.2
moviepy==1.0.3
librosa==0.10.1
python-dotenv==1.0.0
requests==2.31.0
